+++ packages/expojet/CHANGELOG.md
@@ -1,5 +1,12 @@
 # expojet
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies [create-expojet@0.6.2]
+  - create-expojet@0.6.2
+
 ## 0.6.1
 
 ### Patch Changes